refactored to single follower class with selectable control modes
