+++ lerobot_robot_trlc_dk1/__init__.py
@@ -1,11 +1,12 @@
 from .leader import DK1Leader, DK1LeaderConfig
-from .follower import DK1Follower, DK1FollowerConfig
+from .follower import DK1Follower, DK1FollowerConfig, JointImpedanceParams
 from .bi_leader import BiDK1Leader, BiDK1LeaderConfig
 from .bi_follower import BiDK1Follower, BiDK1FollowerConfig
 
 __all__ = [
     "DK1Leader", "DK1LeaderConfig",
     "DK1Follower", "DK1FollowerConfig",
+    "JointImpedanceParams",
     "BiDK1Leader", "BiDK1LeaderConfig",
     "BiDK1Follower", "BiDK1FollowerConfig",
 ]
