+++ lerobot_robot_trlc_dk1/bi_follower.py
@@ -38,6 +38,7 @@ def map_range(x: float, in_min: float, in_max: float, out_min: float, out_max: f
 class BiDK1FollowerConfig(RobotConfig):
     left_arm_port: str
     right_arm_port: str
+    controller_type: str = "pos_vel"  # "pos_vel" | "joint_impedance"
     disable_torque_on_disconnect: bool = False
     joint_velocity_scaling: float = 0.2
     max_gripper_torque: float = 1.0 # Nm (/0.00875m spur gear radius = 114N gripper force)
@@ -59,12 +60,14 @@ class BiDK1Follower(Robot):
         
         left_arm_config = DK1FollowerConfig(
             port=self.config.left_arm_port,
+            controller_type=self.config.controller_type,
             disable_torque_on_disconnect=self.config.disable_torque_on_disconnect,
             joint_velocity_scaling=self.config.joint_velocity_scaling,
             max_gripper_torque=self.config.max_gripper_torque,
         )
         right_arm_config = DK1FollowerConfig(
             port=self.config.right_arm_port,
+            controller_type=self.config.controller_type,
             disable_torque_on_disconnect=self.config.disable_torque_on_disconnect,
             joint_velocity_scaling=self.config.joint_velocity_scaling,
             max_gripper_torque=self.config.max_gripper_torque,
